feat(ui): make Mountable Root (Workspace) router-agnostic (#32)
The Core's top-level Workspace now owns chrome + view switching and
navigates through an injected Router instead of $app/navigation:

- new src/lib/router.svelte.ts: Router contract, RouteTarget, in-memory
  createStateRouter (no URL/history), context setRouter/useRouter, and
  path<->location mapping
- Workspace.svelte repurposed as the Mountable Root (chrome, view switch,
  connection/theme wiring, URL->tab nav-sync); old two-pane moved to
  CollectionsView.svelte; cluster/security pages moved to ClusterView/
  SecurityView lib components
- Topbar/StatusBar/CommandPalette/SchemaTree/ResultsPane/SubNav now read
  and navigate via the router context, not $app/state or goto
- SvelteKit Surface wraps the same root via routes/kit-router.svelte.ts
  (the only $app importer outside route files), keeping URL/deep links;
  route +page.svelte files are URL markers only
- Core ($lib) is now SvelteKit-free

Typecheck 0/0, 179 tests pass, static build OK.
