+++ packages/ui/src/lib/index.ts
@@ -1,5 +1,7 @@
 export { default as ActivityIndicator } from './ActivityIndicator.svelte'
 export { default as ClusterNode } from './ClusterNode.svelte'
+export { default as ClusterView } from './ClusterView.svelte'
+export { default as CollectionsView } from './CollectionsView.svelte'
 export { default as CollectionHistory } from './CollectionHistory.svelte'
 export { default as CommandPalette } from './CommandPalette.svelte'
 export { default as ConnectDropdown } from './ConnectDropdown.svelte'
@@ -11,6 +13,7 @@ export { default as PendingChangesPanel } from './PendingChangesPanel.svelte'
 export { default as QueryEditor } from './QueryEditor.svelte'
 export { default as ResultsPane } from './ResultsPane.svelte'
 export { default as SchemaTree } from './SchemaTree.svelte'
+export { default as SecurityView } from './SecurityView.svelte'
 export { default as ShortcutOverlay } from './ShortcutOverlay.svelte'
 export { default as Splash } from './Splash.svelte'
 export { default as StatusBar } from './StatusBar.svelte'
@@ -28,6 +31,7 @@ export {
 export type { Capability as CollectionCapability } from './capability'
 export * from './collection-catalog'
 export * from './collection-pages'
+export * from './router.svelte'
 export * from './cn'
 export * from './connections.svelte'
 export * from './shortcuts'
