+++ packages/ui/src/routes/collections/+page.svelte
@@ -1,5 +1,5 @@
-<script lang="ts">
-  import Workspace from '$lib/Workspace.svelte'
-</script>
-
-<Workspace />
+<!--
+  URL marker only. The Mountable Root (mounted in +layout.svelte) reads the
+  location from the Kit router and renders the matching view, so this page
+  intentionally renders nothing of its own.
+-->
